requirements.txt
pdf2image
pytesseract
pdfplumber
opencv-python
pillow
jsonlines
fitz
frontend
tools
pymupdf